+++ cmd/wfx/cmd/root/main_test.go
@@ -18,7 +18,6 @@ import (
 const testHost = "localhost"
 
 func TestMain(m *testing.M) {
-	_ = os.Setenv("WFX_CLIENT_HOST", testHost)
-	_ = os.Setenv("WFX_CLIENT_TLS_HOST", testHost)
+	_ = os.Setenv("WFX_CLIENT_HOST", "http://"+testHost+":8080")
 	goleak.VerifyTestMain(m)
 }